workaround bug in parsing library dependencies
The behavior of pyparsing and/or pyfpm changes depending on whether a
token is a subscriptable type (supported in python 3.9). The chisel
base operator class PhysicalOperator inherited (somewhat gratuitously)
from Iterable. As of python 3.9, Iterable is among the standard
collections classes that now support type hinting generics per PEP 585.
A simple workaround turns out to be not inheriting from Iterable. No
functionality is affected based on unit test suite coverage.