Projection supports constraints and mappings
Projecting a relation will also "project" the constraints (keys and
foreign keys) satisfied by the projected attributions. Constraints are
renamed per the DERIVA naming conventions. Then mappings (i.e., source
definitions in schema annotations) that are impacted by attribute and
constraint changes are updated: pruned where the attribute or constraint
are not projected or repaired where the attribute or constraint name are
changed. The assign operation performs a final sweep of the table
definition to finalize constraint names and update mappings accordingly.