Deep copy of metadata by extant physical operators
Before the projection of constraints and mappings, changes to table
definitions could be safely performed by shallow copy of parts of the
definition impacted by the chisel expression. With the introduction of
projection on constraints and mappings, an initial deep copy of the
source table definition by the initial extant physical operator ensures
that manipulations do not alter the catalog's table definition.